AssemblyAnpassung

+++ BankClient/BankClient/CreditAccActions.xaml.cs
@@ -63,7 +63,7 @@ namespace BankClient
 
         private void loeschen(object sender, RoutedEventArgs e)
         {
-            int Kontonummer = global.getAccnumber();
+            long Kontonummer = global.getAccnumber();
             int user = global.getCustID();
             if (Bank.getCreditkontostand(Kontonummer) == 0)
             {
