+++ BankClient/BankClient/DepositAccActions.xaml.cs
@@ -50,7 +50,7 @@ namespace BankClient
 
         private void loeschen(object sender, RoutedEventArgs e)
         {
-            int Kontonummer = global.getAccnumber();
+            long Kontonummer = global.getAccnumber();
             int user = global.getCustID();
             if (Bank.getDepositkontostand(Kontonummer) == 0) {
